chore: added core tests
