+++ src/plugins/todo.rs
@@ -30,7 +30,7 @@ pub struct TodoCli {
 }
 
 #[derive(Subcommand, Debug)]
-enum TodoCommand {
+pub enum TodoCommand {
     /// Add a new task.
     Add {
         #[clap(long)]
@@ -94,23 +94,23 @@ enum TodoCommand {
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-struct Task {
-    id: String,
-    title: String,
-    tags: String,
-    owner: String,
-    due: Option<String>,
-    r#ref: String,
-    status: String,
-    created_at: String,
-    updated_at: String,
-    completed_at: Option<String>,
-    dir_path: String,
-    scope: String,
-    parent_task_id: Option<String>,
-    priority: String,
-    depends_on: String,
-    blocks: String,
+pub struct Task {
+    pub id: String,
+    pub title: String,
+    pub tags: String,
+    pub owner: String,
+    pub due: Option<String>,
+    pub r#ref: String,
+    pub status: String,
+    pub created_at: String,
+    pub updated_at: String,
+    pub completed_at: Option<String>,
+    pub dir_path: String,
+    pub scope: String,
+    pub parent_task_id: Option<String>,
+    pub priority: String,
+    pub depends_on: String,
+    pub blocks: String,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -134,7 +134,7 @@ fn now_iso() -> String {
     format!("{}Z", secs)
 }
 
-fn todo_db_path(root: &Path) -> PathBuf {
+pub fn todo_db_path(root: &Path) -> PathBuf {
     root.join(schemas::TODO_DB_NAME)
 }
 
@@ -264,7 +264,7 @@ fn insert_event(conn: &Connection, ev: &TodoEvent) -> SqlResult<()> {
     Ok(())
 }
 
-fn add_task(root: &Path, args: &TodoCommand) -> Result<serde_json::Value, error::DecapodError> {
+pub fn add_task(root: &Path, args: &TodoCommand) -> Result<serde_json::Value, error::DecapodError> {
     let TodoCommand::Add {
         title,
         tags,
@@ -356,7 +356,7 @@ fn add_task(root: &Path, args: &TodoCommand) -> Result<serde_json::Value, error:
     }))
 }
 
-fn update_status(
+pub fn update_status(
     store: &Store,
     id: &str,
     new_status: &str,
@@ -448,7 +448,7 @@ fn comment_task(
     }))
 }
 
-fn get_task(root: &Path, id: &str) -> Result<Option<Task>, error::DecapodError> {
+pub fn get_task(root: &Path, id: &str) -> Result<Option<Task>, error::DecapodError> {
     let broker = DbBroker::new(root);
     let db_path = todo_db_path(root);
 
@@ -481,7 +481,7 @@ fn get_task(root: &Path, id: &str) -> Result<Option<Task>, error::DecapodError> 
     })
 }
 
-fn list_tasks(
+pub fn list_tasks(
     root: &Path,
     status: Option<String>,
     scope: Option<String>,
@@ -557,7 +557,7 @@ fn list_tasks(
     })
 }
 
-fn rebuild_from_events(root: &Path) -> Result<serde_json::Value, error::DecapodError> {
+pub fn rebuild_from_events(root: &Path) -> Result<serde_json::Value, error::DecapodError> {
     let ev_path = events_path(root);
     if !ev_path.is_file() {
         // Empty store is valid; create empty DB with schema.
@@ -826,86 +826,4 @@ pub fn run_todo_cli(store: &Store, cli: TodoCli) -> Result<(), error::DecapodErr
     Ok(())
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
 
-    #[test]
-    fn test_todo_lifecycle() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        initialize_todo_db(&root).unwrap();
-
-        // 1. Add task
-        let add_args = TodoCommand::Add {
-            title: "Test task".to_string(),
-            tags: "tag1".to_string(),
-            owner: "arx".to_string(),
-            due: None,
-            r#ref: "".to_string(),
-            dir: Some(tmp.path().to_string_lossy().to_string()),
-            priority: "high".to_string(),
-            depends_on: "".to_string(),
-            blocks: "".to_string(),
-            parent: None,
-        };
-        let res = add_task(&root, &add_args).unwrap();
-        let task_id = res.get("id").unwrap().as_str().unwrap();
-        assert!(task_id.contains("_"));
-
-        // 2. Get task
-        let task = get_task(&root, task_id).unwrap().expect("Task not found");
-        assert_eq!(task.title, "Test task");
-        assert_eq!(task.status, "open");
-
-        // 3. Mark done
-        let store = Store {
-            kind: crate::core::store::StoreKind::Repo,
-            root: root.clone(),
-        };
-        update_status(&store, task_id, "done", "task.done", serde_json::json!({})).unwrap();
-        let task = get_task(&root, task_id).unwrap().unwrap();
-        assert_eq!(task.status, "done");
-
-        // 4. List tasks
-        let tasks = list_tasks(&root, Some("done".to_string()), None, None, None, None).unwrap();
-        assert_eq!(tasks.len(), 1);
-        assert_eq!(tasks[0].id, task_id);
-    }
-
-    #[test]
-    fn test_todo_rebuild() {
-        let tmp = tempdir().unwrap();
-        let root = tmp.path().to_path_buf();
-        initialize_todo_db(&root).unwrap();
-
-        // Add some tasks
-        for i in 0..3 {
-            let add_args = TodoCommand::Add {
-                title: format!("Task {}", i),
-                tags: "".to_string(),
-                owner: "".to_string(),
-                due: None,
-                r#ref: "".to_string(),
-                dir: Some(tmp.path().to_string_lossy().to_string()),
-                priority: "medium".to_string(),
-                depends_on: "".to_string(),
-                blocks: "".to_string(),
-                parent: None,
-            };
-            add_task(&root, &add_args).unwrap();
-        }
-
-        // Corrupt/Delete DB
-        let db_path = todo_db_path(&root);
-        fs::remove_file(&db_path).unwrap();
-
-        // Rebuild
-        rebuild_from_events(&root).unwrap();
-
-        // Verify
-        let tasks = list_tasks(&root, None, None, None, None, None).unwrap();
-        assert_eq!(tasks.len(), 3);
-    }
-}
