+++ tests/plugins/mod.rs
@@ -1,0 +1,2 @@
+pub mod todo;
+pub mod policy;
